Test(store): RegionCode - String 타입 변환, 최소 주문 금액 int 설정 값 Store Entity에서 minOrderAmount int -> Long type 변환 후 fix

+++ src/main/java/app/domain/store/model/entity/Store.java
@@ -62,7 +62,7 @@ public class Store extends BaseEntity {
 	private String phoneNumber;
 
 	@Column(nullable = false)
-	private int minOrderAmount = 0;
+	private long minOrderAmount = 0;
 
 	@Enumerated(EnumType.STRING)
 	@Column(nullable = false)
